### 2019-08-02 #1623 @labarba
please merge this fix
https://github.com/richteague/gofish/pull/2
