### 2019-08-02 #1623 @labarba
@arfon — the submitting author suggested you as handling editor. Could you take it?

**GoFish: Fishing for Line Observations in Protoplanetary Disks**
https://github.com/richteague/gofish
[ Article proof :page_facing_up: ](https://github.com/openjournals/joss-papers/blob/joss.01623/joss.01623/10.21105.joss.01623.pdf)
