### 2019-08-02 #1623 @arfon
:wave: @richteague, thanks for your submission to JOSS! A couple of quick questions:

- Is this paper accompanying a submission in a AAS publication?
- Could you take a look a [this list of potential reviewers](https://bit.ly/joss-reviewers) and identify a few people who would be good candidates to review this submission?
