### 2019-08-05 #1623 @richteague
Hi @arfon!

No, this paper should be a stand alone, although the methods have previously been described in various AAS articles.

For reviewers, I could suggest ygrange, nespinoza or zhampel, although I'd be happy with any reviewer that's familiar with astronomy.
