### 2019-08-07 #1623 @arfon
> No, this paper should be a stand alone, although the methods have previously been described in various AAS articles.

👍understood. Thanks.

:wave: @zhampel & @nespinoza - would you be willing to review this submission for JOSS?
