### 2019-08-07 #1623 @richteague
So I'm actually on vacation for the next two and a bit weeks so maybe this would work? I'll be back in the office early September so from my side there's no rush.
