### 2019-08-08 #1623 @arfon
> If that's the case then yes, I could start reviewing this the first days of September :-). Does this sound good @arfon?

Sounds great to me!
