### 2019-08-08 #1632 @arfon
Re-stating here that @nespinoza won't be able to start his review until early September.
