### 2019-08-09 #1632 @zhampel
@richteague

*Paper*
Corrections/Clarifications:
- requirer -> require
- which is assumed to random -? which is assumed to be random / which is assumed random
- Can you please specify what is meant by random, i.e. type or distribution?
- Does 'correcting' need to be in quotes? It is a correction operation itself, no?
- mass of the mass of the central star -> mass of the central star

*Documentation*

- I recommend you provide a clearly labelled section on your GitHub README that provides some basic form of documentation & usage. Preferably also a clear link in this section that directs to your doc page. The little icon below the molecular fishing image is a bit small and easily bypassed.
- Your readthedocs page is nice & thorough.
- The Harvard Dataverse link directs to a page claiming that the service is unavailable. Is this a bad link or an intermittent issue?

Once access is permitted to the example data from Harvard Dataverse, I will continue my review of the codebase operability/functionality.
