### 2019-09-04 #1632 @richteague
@zhampel

Thanks for the initial comments!

I've corrected the manuscript draft following your suggestions. I've also updated the README to include a brief example usage and another link to the documentation. Please let me know if you feel as if there should be more here.

Regarding the data files, these should be persistent and are likely an intermittent issue. I have checked the downloads and had others verify that they're able to download the data cube so hopefully everything is good to go.
