### 2019-09-05 #1632 @nespinoza
Hi everyone!

First of all, congrats @richteague on this very useful piece of software! I tested the whole code and all worked for me (in particular, the TWHya dataset worked for me). I have gone over the documentation and paper, and have a few minor comments I would like to cover, and one moderate comment at the end:

**Regarding the paper**:

- For some reason the references are not compiling correctly --- I think this has to do with the fact that the `paper.md` file in the repository is performing the citations using commas between each citation, i.e., you are doing `'[@citation]'` instead of simply doing `[@citation]`.

**Regarding the code & documentation**:

- In the documentation, I had trouble viewing line [10] of the notebook (Fishing Basics --- "Extracing a Disk Averaged Spectrum") in my computer (I had to go to the actual notebook to run/see this line in my own system). This might be because the line is too long --- perhaps you could consider breaking the line in two with `\`.
- A bit of a technical question: in line [12] and [13] of the notebook of the same section, you play with resampling the data. However, the errorbars go _up_ when you bin the data in velocity-space where they should intuitively go _down_ (you have more data on each bin, hence, errors on the flux density should go down as ~`1/sqrt(n)` where `n` is the number of samples in the bin). I checked the code and it seems this goes back to `eddy`, which in turn calls `scipy.stats.binned_statistics`. Turns out that when you ask for `std` in `scipy.stats.binned_statistics` this returns the _sample_ standard deviation on each bin, whereas what you really want, I believe, is the _error on the flux density_ for the given bin. Assuming independent and identically distributed data, a good measure of this is the standard error on the mean, which is given by the sample standard deviation divided by the square-root of `n`. I believe it is important this is reported in the documentation, because some people (like me) would perhaps be expecting the error on the flux densities to be reported and not the sample standard deviation of the bin, which are different statistics.
