### 2019-09-05 81f9fde
updated uncertainty calculation
