### 2019-09-05 #1632 @richteague
Hi @nespinoza!

Thanks for the comments!

I've corrected the citation style, hopefully this works now.

I've included a line break in line 10 of the Fishing Basics notebook and appears to render OK.

I've added a small community guidelines section pointing towards the (albeit brief) contributing guide on the main repo. If you think this require more info I can update this.

Regarding the technical question, you're absolutely right. I made the choice in `eddy` to use just the standard deviation because the spectra will have some small levels of correlation (both spatially and spectrally). However, I've included a new argument, `assume_correlated` which, if set to `False` (the default is `True`), it estimates the error following the poisson statistics you described above. I've included both a note in the function docstrings to highlight this and included another cell in the notebook demonstrating this.
