### 2019-09-05 #1632 @nespinoza
Thanks for the quick response to the comments @richteague! I'm all set then --- ready to accept the paper to JOSS (as soon as the code has a release version in Github, so I can mark the "version" part above)!
