### 2019-09-10 #1632 @zhampel
@richteague

Thanks for your changes, and after having been able to download the dataset, I see that the functionality is as advertised. BTW the documentation is very thorough and appreciated!

However, I have a request regarding the installation which can have an impact on those wishing to contribute (e.g. if users fork the repo and wish to make PRs). I recommend that you add an `__init__.py` file in the gofish with the following lines:

# __init__.py

from .gofish import *

This would allow a user to (instead of running `pip install gofish`) clone the repo, setup a virtualenv and run `python setup.py install`. Then the user would have an easier time testing and contributing to your codebase. After such a change is made, I will check off the installation & contributing boxes, as this will facilitate user contributions. Then I think you can update the version. Is that acceptable?

One other thing: have you considered adding an emcee search to the advanced fishing example?
