+++ gofish/__init__.py
@@ -1,0 +1,4 @@
+# -*- coding: utf-8 -*-
+
+from .gofish import imagecube
+__all__ = ["imagecube"]
