### 2019-09-10 #1632 @richteague
@zhampel - Thanks a lot for the report! I've added the `__init__.py` file to the main repo. I've thought about adding an emcee search to the documentation and this is something I want to to, however I'm working with a collaborator to get some real data that would be more appropriate for such an example. As it might take a little while to get the data, I didn't want that to delay the publication of the software.

I've also updated the version to v1.0 in GitHub (@nespinoza).

Thanks again to you both for your quick reviews!
