### 2019-09-10 #1632 @arfon
@richteague - minor point, could you remove Lorena's name from the authors of the Zenodo archive?

Also, could you update the title of your submission to match the paper, i.e. change it to `GoFish: Fishing for Line Observations in Protoplanetary Disks`
