### 2019-09-13 #1632 @Kevin-Mattheus-Moerman
@arfon this paper lacks a **References** heading. Should one be added?
