### 2019-09-15 #1632 @arfon
> @arfon this paper lacks a References heading. Should one be added?

Good catch @Kevin-Mattheus-Moerman.

@richteague - could you please merge this small change to your paper? https://github.com/richteague/gofish/compare/master...arfon:patch-1
